Fixed background colors in ProfileComponent and Contact.module.scss Added FullWidthBackground component for consistent styling. Added HomePage
