+++ src/components/ProfileComponent/FullWidthBackground.css
@@ -1,0 +1,11 @@
+.full-width-bg {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background: #1d2026;
+  margin: 0;
+  padding: 0;
+  z-index: -1;
+}
