+++ src/components/ProfileComponent/FullWidthBackground.js
@@ -1,0 +1,10 @@
+import React from "react";
+import "./FullWidthBackground.css";
+
+const FullWidthBackground = ({ children }) => (
+  <div className="full-width-bg">
+    {children}
+  </div>
+);
+
+export default FullWidthBackground;
