+++ src/components/ProfileComponent/Hero.module.scss
