+++ src/components/ProfileComponent/HomePage.module.scss
@@ -1,0 +1,27 @@
+.homePage {
+  min-height: 100vh;
+  width: 100vw;
+  background: linear-gradient(135deg, #1e2127 0%, #505764 100%);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+.homeHeader {
+  background: rgba(40,44,52,0.95);
+  border-radius: 16px;
+  padding: 2rem 3rem;
+  box-shadow: 0 4px 24px rgba(0,0,0,0.10);
+  color: white;
+  width: 100%;
+  max-width: 600px;
+  margin: 2rem auto;
+  text-align: center;
+}
+.homeLogo {
+  height: 40vmin;
+  pointer-events: none;
+}
+.homeLink {
+  color: #61dafb;
+}
